Code cleanup. Removed the token refresh logics. Removed validating JWT Token and related components.
